Reorganize daphne worker
